rework solid for web_hook
