+++ src/core/aiohttp_web_hook.py
@@ -1,7 +1,12 @@
 import asyncio
+
 from aiohttp import web
-import telebot
 from aiohttp.web import Request
+
+import telebot
+from telebot.async_telebot import AsyncTeleBot
+
+from core.config import SettingsBot
 
 
 async def handle(request: Request):
@@ -13,3 +18,19 @@ async def handle(request: Request):
         return web.Response()
     else:
         return web.Response(status=403)
+
+
+async def shutdown(app):
+    bot_instance = app['bot_instance']
+    await bot_instance.remove_webhook()
+    await bot_instance.close_session()
+
+
+async def setup(bot_instance: AsyncTeleBot, settings: SettingsBot):
+    # Remove webhook, it fails sometimes the set if there is a previous webhook
+    app = web.Application()
+    app.router.add_post('/{token}/', handle)
+    app.on_cleanup.append(shutdown)
+    app['bot_instance'] = bot_instance
+    app['bot_settings'] = settings
+    return app
