+++ src/main.py
@@ -8,7 +8,7 @@ from telebot.async_telebot import AsyncTeleBot
 from core.bot import bot
 from core.log_config import loger
 from core.config import SettingsBot, bot_settings
-from core.aiohttp_web_hook import handle
+from core.aiohttp_web_hook import setup
 
 from custom_handlers.group.message import register_chat_custom_message_handlers
 from custom_handlers.private.message import register_custom_message_handlers
@@ -39,45 +39,28 @@ async def register_handlers(bot: AsyncTeleBot) -> None:
     )
 
 
-async def registration(bot: AsyncTeleBot, loger: logging, settings: SettingsBot):
+async def registration(bot_instance: AsyncTeleBot, loger: logging, settings: SettingsBot):
     loger.info("Starting bot")
-    # await register_middleware(bot)
-    await register_log_middleware(bot, loger)
-    await register_message_filters(bot, loger)
-    await register_callback_filters(bot, loger)
-    await register_handlers(bot)
+    # await register_middleware(bot_instance)
+    await register_log_middleware(bot_instance, loger)
+    await register_message_filters(bot_instance, loger)
+    await register_callback_filters(bot_instance, loger)
+    await register_handlers(bot_instance)
     loger.info('Starting up: removing old webhook')
-    await bot.remove_webhook()
+    await bot_instance.remove_webhook()
     # Set webhook
     loger.info('Starting up: setting webhook')
-    await bot.set_webhook(
+    await bot_instance.set_webhook(
         url=settings.webhook_url_base.format(settings.webhook_host, settings.webhook_port)
         + settings.webhook_url_path.format(settings.bot_token),
         certificate=open(settings.webhook_ssl_cert, 'r')
     )
 
 
-async def shutdown(app):
-    loger.info('Shutting down: removing webhook')
-    bot_instance = app['bot_instance']
-    await bot_instance.remove_webhook()
-    loger.info('Shutting down: closing session')
-    await bot_instance.close_session()
-
-
-async def setup(bot_instance: AsyncTeleBot, settings: SettingsBot):
-    # Remove webhook, it fails sometimes the set if there is a previous webhook
-    app = web.Application()
-    app.router.add_post('/{token}/', handle)
-    app.on_cleanup.append(shutdown)
-    app['bot_instance'] = bot_instance
-    app['bot_settings'] = settings
-    return app
-
 if __name__ == '__main__':
     context = ssl.SSLContext(ssl.PROTOCOL_TLSv1_2)
     context.load_cert_chain(bot_settings.webhook_ssl_cert, bot_settings.webhook_ssl_priv)
-    asyncio.run(registration(bot=bot, loger=loger, settings=bot_settings))
+    asyncio.run(registration(bot_instance=bot, loger=loger, settings=bot_settings))
     web.run_app(
         setup(bot_instance=bot, settings=bot_settings),
         host=bot_settings.webhook_listen,
